Split new_block related logic to another file
